[#52] improved pagination for products

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/ProductService.java
@@ -62,7 +62,7 @@ public interface ProductService {
      *
      * @return retrieves the number of all added products
      */
-    int getProductsCount();
+    Long getProductsCount();
 
     /**
      * Deletes a specific product with the given productId.
